requirements.txt
discord.py==2.3.2
anthropic==0.25.0
gspread==6.1.2
google-auth==2.29.0
